requirements.txt

transformers==4.30.2
sentence-transformers==2.2.2
huggingface_hub==0.14.1
torch==2.2.2
accelerate==0.29.3
bitsandbytes==0.43.0
einops==0.8.0
langchain==0.2.0
nltk==3.8.1
PyPDF2==3.0.1
faiss-cpu==1.7.4
